+++ android/settings.gradle
@@ -1,1 +1,1 @@
-rootProject.name = 'contacts_ui_service'
+rootProject.name = 'native_contact_dialog'
