--- ios/Classes/ContactsUiServicePlugin.h
@@ -1,4 +1,0 @@
-#import <Flutter/Flutter.h>
-
-@interface ContactsUiServicePlugin : NSObject<FlutterPlugin>
-@end
