--- ios/Classes/ContactsUiServicePlugin.m
@@ -1,9 +1,0 @@
-#import "ContactsUiServicePlugin.h"
-#import <contacts_ui_service/contacts_ui_service-Swift.h>
-
-@implementation ContactsUiServicePlugin
-+ (void)registerWithRegistrar:(NSObject<FlutterPluginRegistrar>*)registrar {
-  [SwiftContactsUiServicePlugin registerWithRegistrar:registrar];
-}
-  
-@end
