+++ ios/Classes/NativeContactDialogPlugin.h
@@ -1,0 +1,4 @@
+#import <Flutter/Flutter.h>
+
+@interface NativeContactDialogPlugin : NSObject<FlutterPlugin>
+@end
