+++ pubspec.yaml
@@ -1,7 +1,8 @@
-name: contacts_ui_service
-description: A new flutter plugin project.
+name: native_contact_dialog
+description: A Flutter plugin to add new contacts using the platform's native contact dialogs.
 version: 0.0.1
-author:
+auathor:
+- Andrew Fulton <afuton11@gmail.com>
 homepage:
 
 environment:
@@ -21,8 +22,8 @@ flutter:
   # be modified. They are used by the tooling to maintain consistency when
   # adding or updating assets for this project.
   plugin:
-    androidPackage: com.example.contactsuiservice
-    pluginClass: ContactsUiServicePlugin
+    androidPackage: com.fulton.nativecontactdialog
+    pluginClass: NativeContactDialogPlugin
 
   # To add assets to your plugin package, add an assets section, like this:
   # assets:
